+++ src/main/java/com/gladkiei/exchanger/dto/ExchangeRateRequestDTO.java
@@ -3,14 +3,16 @@ package com.gladkiei.exchanger.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Setter
 @Getter
 public class ExchangeRateRequestDTO {
     private String baseCurrencyCode;
     private String targetCurrencyCode;
-    private double rate;
+    private BigDecimal rate;
 
-    public ExchangeRateRequestDTO(String baseCurrencyCode, String targetCurrencyCode, double rate) {
+    public ExchangeRateRequestDTO(String baseCurrencyCode, String targetCurrencyCode, BigDecimal rate) {
         this.baseCurrencyCode = baseCurrencyCode;
         this.targetCurrencyCode = targetCurrencyCode;
         this.rate = rate;
