+++ src/main/java/com/gladkiei/exchanger/models/ExchangeRate.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -11,5 +13,5 @@ public class ExchangeRate {
     private int id;
     private Currency baseCurrency;
     private Currency targetCurrency;
-    double rate;
+    BigDecimal rate;
 }
